Isolate execution contexts for a chain of dependent executions. Simplified ReplForJupyterImpl. Extracted InternalEvaluator and CellExecutor interfaces. Eliminated PreprocessingResult class. Redesigned interaction between Execution and KotlinKernelHost. Extracted SharedReplContext with a bunch of components that are used in CellExecutor. Changed LibraryResolver to return LibraryDefinition instead of LibraryDescriptor to support callback-based LibraryDefinitions (primarily, in tests).
