+++ jupyter-lib/api/src/main/kotlin/org/jetbrains/kotlinx/jupyter/api/libraries/CodeExecution.kt
@@ -13,12 +13,20 @@ import org.jetbrains.kotlinx.jupyter.util.replaceVariables
  * @constructor Create code execution with the given [code]
  */
 @Serializable(with = CodeExecutionSerializer::class)
-class CodeExecution(val code: Code) : Execution {
-    override fun execute(host: KotlinKernelHost): Any? {
-        return host.execute(code)
+class CodeExecution(val code: Code) : Execution<Any?> {
+    override fun execute(host: ExecutionHost): Any? {
+        return host.execute {
+            execute(code)
+        }
     }
 
     override fun replaceVariables(mapping: Map<String, String>): CodeExecution {
         return CodeExecution(replaceVariables(code, mapping))
+    }
+}
+
+class DelegatedExecution<T>(private val callback: KotlinKernelHost.() -> T) : Execution<T> {
+    override fun execute(host: ExecutionHost): T {
+        return host.execute(callback)
     }
 }
