+++ jupyter-lib/api/src/main/kotlin/org/jetbrains/kotlinx/jupyter/api/libraries/Execution.kt
@@ -1,18 +1,16 @@
 package org.jetbrains.kotlinx.jupyter.api.libraries
 
-import org.jetbrains.kotlinx.jupyter.api.KotlinKernelHost
-
 /**
  * Logic to be executed with Kernel REPL
  */
-fun interface Execution : VariablesSubstitutionAware<Execution> {
+fun interface Execution<T> : VariablesSubstitutionAware<Execution<T>> {
     /**
      * Execute itself with a given [host]
      *
      * @param host Kernel host to be executed on
      * @return Execution result
      */
-    fun execute(host: KotlinKernelHost): Any?
+    fun execute(host: ExecutionHost): T
 
-    override fun replaceVariables(mapping: Map<String, String>): Execution = this
+    override fun replaceVariables(mapping: Map<String, String>): Execution<T> = this
 }
