+++ jupyter-lib/api/src/main/kotlin/org/jetbrains/kotlinx/jupyter/api/libraries/ExecutionHost.kt
@@ -1,0 +1,13 @@
+package org.jetbrains.kotlinx.jupyter.api.libraries
+
+import org.jetbrains.kotlinx.jupyter.api.KotlinKernelHost
+
+/**
+ * Used in [Execution.execute] to get [KotlinKernelHost] context
+ */
+interface ExecutionHost {
+
+    fun <T> execute(callback: KotlinKernelHost.() -> T): T
+
+    fun <T> execute(execution: Execution<T>) = execution.execute(this)
+}
