+++ jupyter-lib/api/src/main/kotlin/org/jetbrains/kotlinx/jupyter/api/libraries/LibraryDefinitionImpl.kt
@@ -2,6 +2,7 @@ package org.jetbrains.kotlinx.jupyter.api.libraries
 
 import org.jetbrains.kotlinx.jupyter.api.AnnotationHandler
 import org.jetbrains.kotlinx.jupyter.api.GenerativeTypeHandler
+import org.jetbrains.kotlinx.jupyter.api.KotlinKernelVersion
 import org.jetbrains.kotlinx.jupyter.api.RendererTypeHandler
 
 /**
@@ -13,11 +14,12 @@ class LibraryDefinitionImpl(
     override val dependencies: List<String> = emptyList(),
     override val repositories: List<String> = emptyList(),
     override val imports: List<String> = emptyList(),
-    override val init: List<Execution> = emptyList(),
-    override val initCell: List<Execution> = emptyList(),
-    override val shutdown: List<Execution> = emptyList(),
+    override val init: List<Execution<*>> = emptyList(),
+    override val initCell: List<Execution<*>> = emptyList(),
+    override val shutdown: List<Execution<*>> = emptyList(),
     override val renderers: List<RendererTypeHandler> = emptyList(),
     override val converters: List<GenerativeTypeHandler> = emptyList(),
     override val annotations: List<AnnotationHandler> = emptyList(),
     override val resources: List<LibraryResource> = emptyList(),
+    override val minKernelVersion: KotlinKernelVersion? = null,
 ) : LibraryDefinition
