+++ jupyter-lib/lib/src/main/kotlin/jupyter/kotlin/KotlinKernelHostProvider.kt
@@ -1,0 +1,7 @@
+package jupyter.kotlin
+
+import org.jetbrains.kotlinx.jupyter.api.KotlinKernelHost
+
+interface KotlinKernelHostProvider {
+    val host: KotlinKernelHost?
+}
