+++ jupyter-lib/shared-compiler/src/main/kotlin/org/jetbrains/kotlinx/jupyter/libraries/LibrariesProcessor.kt
@@ -1,69 +1,20 @@
 package org.jetbrains.kotlinx.jupyter.libraries
 
 import org.jetbrains.kotlinx.jupyter.api.KotlinKernelVersion
-import org.jetbrains.kotlinx.jupyter.api.libraries.LibraryDefinitionImpl
+import org.jetbrains.kotlinx.jupyter.api.libraries.LibraryDefinition
 import org.jetbrains.kotlinx.jupyter.api.libraries.LibraryDefinitionProducer
 import org.jetbrains.kotlinx.jupyter.compiler.util.ReplCompilerException
-import org.jetbrains.kotlinx.jupyter.util.replaceVariables
 
-class LibrariesProcessor(
+interface LibrariesProcessor {
+    fun processNewLibraries(arg: String): List<LibraryDefinitionProducer>
+    val libraryFactory: LibraryFactory
+}
+
+class LibrariesProcessorImpl(
     private val libraries: LibraryResolver?,
     private val kernelVersion: KotlinKernelVersion?,
-    val libraryFactory: LibraryFactory,
-) {
-
-    /**
-     * Matches a list of actual library arguments with declared library parameters
-     * Arguments can be named or not. Named arguments should be placed after unnamed
-     * Parameters may have default value
-     *
-     * @return A name-to-value map of library arguments
-     */
-    private fun substituteArguments(parameters: List<Variable>, arguments: List<Variable>): Map<String, String> {
-        val result = mutableMapOf<String, String>()
-        if (arguments.any { it.name.isEmpty() }) {
-            if (parameters.count() != 1) {
-                throw ReplCompilerException("Unnamed argument is allowed only if library has a single property")
-            }
-            if (arguments.count() != 1) {
-                throw ReplCompilerException("Too many arguments")
-            }
-            result[parameters[0].name] = arguments[0].value
-            return result
-        }
-
-        arguments.forEach {
-            result[it.name] = it.value
-        }
-        parameters.forEach {
-            if (!result.containsKey(it.name)) {
-                result[it.name] = it.value
-            }
-        }
-        return result
-    }
-
-    private fun processDescriptor(library: LibraryDescriptor, mapping: Map<String, String>): LibraryDefinitionProducer {
-        val definitionCodes = library.libraryDefinitions
-        return if (definitionCodes.isEmpty()) {
-            TrivialLibraryDefinitionProducer(
-                LibraryDefinitionImpl(
-                    dependencies = library.dependencies.replaceVariables(mapping),
-                    repositories = library.repositories.replaceVariables(mapping),
-                    imports = library.imports.replaceVariables(mapping),
-                    init = library.init.replaceVariables(mapping),
-                    shutdown = library.shutdown.replaceVariables(mapping),
-                    initCell = library.initCell.replaceVariables(mapping),
-                    renderers = library.renderers.replaceVariables(mapping),
-                    converters = library.converters.replaceVariables(mapping),
-                    resources = library.resources.replaceVariables(mapping),
-                )
-            )
-        } else {
-            val initCodes = library.buildDependenciesInitCode(mapping)?.let { listOf(it) } ?: emptyList()
-            ResolvingLibraryDefinitionProducer(initCodes, definitionCodes)
-        }
-    }
+    override val libraryFactory: LibraryFactory,
+) : LibrariesProcessor {
 
     /**
      * Split a command argument into a set of library calls
@@ -100,10 +51,8 @@ class LibrariesProcessor(
         }
     }
 
-    private fun checkKernelVersionRequirements(name: String, library: LibraryDescriptor) {
-        library.minKernelVersion?.let { minVersionStr ->
-            val minVersion = KotlinKernelVersion.from(minVersionStr)
-                ?: throw ReplCompilerException("Wrong format of minimal kernel version for library '$name': $minVersionStr")
+    private fun checkKernelVersionRequirements(name: String, library: LibraryDefinition) {
+        library.minKernelVersion?.let { minVersion ->
             kernelVersion?.let { currentVersion ->
                 if (currentVersion < minVersion) {
                     throw ReplCompilerException("Library '$name' requires at least $minVersion version of kernel. Current kernel version is $currentVersion. Please update kernel")
@@ -112,15 +61,14 @@ class LibrariesProcessor(
         }
     }
 
-    fun processNewLibraries(arg: String) =
+    override fun processNewLibraries(arg: String): List<LibraryDefinitionProducer> =
         splitLibraryCalls(arg).map {
             val (libRef, vars) = libraryFactory.parseReferenceWithArgs(it)
-            val library = libraries?.resolve(libRef)
+            val library = libraries?.resolve(libRef, vars)
                 ?: throw ReplCompilerException("Unknown library '$libRef'")
+
             checkKernelVersionRequirements(libRef.toString(), library)
 
-            val mapping = substituteArguments(library.variables, vars)
-
-            processDescriptor(library, mapping)
+            TrivialLibraryDefinitionProducer(library)
         }
 }
