+++ src/main/kotlin/org/jetbrains/kotlinx/jupyter/codegen/TypeRenderersProcessorImpl.kt
@@ -1,31 +1,30 @@
 package org.jetbrains.kotlinx.jupyter.codegen
 
 import org.jetbrains.kotlinx.jupyter.api.Code
-import org.jetbrains.kotlinx.jupyter.api.KotlinKernelHost
 import org.jetbrains.kotlinx.jupyter.api.PrecompiledRendererTypeHandler
 import org.jetbrains.kotlinx.jupyter.api.RendererTypeHandler
+import org.jetbrains.kotlinx.jupyter.api.libraries.ExecutionHost
 import org.jetbrains.kotlinx.jupyter.repl.ContextUpdater
 
 class TypeRenderersProcessorImpl(
-    private val host: KotlinKernelHost,
     private val contextUpdater: ContextUpdater,
 ) : TypeRenderersProcessor {
     private var counter = 0
     private val typeRenderers: MutableList<HandlerWithInfo> = mutableListOf()
 
-    override tailrec fun renderResult(value: Any?, fieldName: String?): Any? {
+    override tailrec fun renderResult(host: ExecutionHost, value: Any?, fieldName: String?): Any? {
         if (value == null) return null
         val (handler, id) = typeRenderers.firstOrNull { it.handler.acceptsType(value::class) }
             ?: return value
         return if (id == null) {
             val (resultValue, resultFieldName) = handler.execution.execute(host, value, fieldName)
-            renderResult(resultValue, resultFieldName)
+            renderResult(host, resultValue, resultFieldName)
         } else {
             val methodName = getMethodName(id)
             contextUpdater.update()
             val functionInfo = contextUpdater.context.functions[methodName]!!
             val resultValue = functionInfo.function.call(functionInfo.line, value)
-            renderResult(resultValue, null)
+            renderResult(host, resultValue, null)
         }
     }
