+++ src/main/kotlin/org/jetbrains/kotlinx/jupyter/repl/CellExecutor.kt
@@ -1,0 +1,23 @@
+package org.jetbrains.kotlinx.jupyter.repl
+
+import org.jetbrains.kotlinx.jupyter.DisplayHandler
+import org.jetbrains.kotlinx.jupyter.api.Code
+import org.jetbrains.kotlinx.jupyter.api.libraries.ExecutionHost
+
+/**
+ * Executes notebook cell code.
+ * Performs code preprocessing (magics parsing) and snippet class postprocessing (variable converters and annotation handlers)
+ */
+interface CellExecutor : ExecutionHost {
+
+    fun execute(
+        code: Code,
+        displayHandler: DisplayHandler? = null,
+        processVariables: Boolean = true,
+        processAnnotations: Boolean = true,
+        processMagics: Boolean = true,
+        callback: ExecutionStartedCallback? = null
+    ): InternalEvalResult
+}
+
+typealias ExecutionStartedCallback = (internalId: Int, code: Code) -> Unit
