+++ src/main/kotlin/org/jetbrains/kotlinx/jupyter/repl/impl/InternalEvaluatorImpl.kt
@@ -1,0 +1,107 @@
+package org.jetbrains.kotlinx.jupyter.repl.impl
+
+import kotlinx.coroutines.runBlocking
+import org.jetbrains.kotlinx.jupyter.ReplEvalRuntimeException
+import org.jetbrains.kotlinx.jupyter.api.Code
+import org.jetbrains.kotlinx.jupyter.compiler.CompiledScriptsSerializer
+import org.jetbrains.kotlinx.jupyter.compiler.util.ReplCompilerException
+import org.jetbrains.kotlinx.jupyter.compiler.util.SourceCodeImpl
+import org.jetbrains.kotlinx.jupyter.repl.ContextUpdater
+import org.jetbrains.kotlinx.jupyter.repl.InternalEvalResult
+import org.jetbrains.kotlinx.jupyter.repl.InternalEvaluator
+import kotlin.script.experimental.api.ResultValue
+import kotlin.script.experimental.api.ResultWithDiagnostics
+import kotlin.script.experimental.jvm.BasicJvmReplEvaluator
+
+internal class InternalEvaluatorImpl(val compiler: JupyterCompiler<*>, val evaluator: BasicJvmReplEvaluator, val contextUpdater: ContextUpdater, override var logExecution: Boolean) :
+    InternalEvaluator {
+
+    private var classWriter: ClassWriter? = null
+
+    private val scriptsSerializer = CompiledScriptsSerializer()
+
+    override var writeCompiledClasses: Boolean
+        get() = classWriter != null
+        set(value) {
+            classWriter = if (!value) null
+            else {
+                val cw = ClassWriter()
+                System.setProperty("spark.repl.class.outputDir", cw.outputDir.toString())
+                cw
+            }
+        }
+
+    override val lastKClass get() = compiler.lastKClass
+
+    override val lastClassLoader get() = compiler.lastClassLoader
+
+    private var isExecuting = false
+
+    override fun eval(code: Code, onInternalIdGenerated: ((Int) -> Unit)?): InternalEvalResult {
+        try {
+            if (isExecuting) {
+                error("Recursive execution is not supported")
+            }
+
+            isExecuting = true
+            if (logExecution) {
+                println("Executing:\n" + code)
+            }
+            val id = compiler.nextCounter()
+
+            if (onInternalIdGenerated != null) {
+                onInternalIdGenerated(id)
+            }
+
+            val codeLine = SourceCodeImpl(id, code)
+
+            val (compileResult, evalConfig) = compiler.compileSync(codeLine)
+            val compiledScript = compileResult.get()
+
+            classWriter?.writeClasses(codeLine, compiledScript)
+            val resultWithDiagnostics = runBlocking { evaluator.eval(compileResult, evalConfig) }
+            contextUpdater.update()
+
+            when (resultWithDiagnostics) {
+                is ResultWithDiagnostics.Success -> {
+                    val pureResult = resultWithDiagnostics.value.get()
+                    return when (val resultValue = pureResult.result) {
+                        is ResultValue.Error -> throw ReplEvalRuntimeException(
+                            resultValue.error.message.orEmpty(),
+                            resultValue.error
+                        )
+                        is ResultValue.Unit -> {
+                            InternalEvalResult(Unit, null, scriptsSerializer.serialize(compiledScript))
+                        }
+                        is ResultValue.Value -> {
+                            InternalEvalResult(
+                                resultValue.value,
+                                pureResult.compiledSnippet.resultField?.first,
+                                scriptsSerializer.serialize(compiledScript)
+                            )
+                        }
+                        is ResultValue.NotEvaluated -> {
+                            throw ReplEvalRuntimeException(
+                                buildString {
+                                    val cause = resultWithDiagnostics.reports.firstOrNull()?.exception
+                                    val stackTrace = cause?.stackTrace.orEmpty()
+                                    append("This snippet was not evaluated: ")
+                                    appendLine(cause.toString())
+                                    for (s in stackTrace)
+                                        appendLine(s)
+                                }
+                            )
+                        }
+                        else -> throw IllegalStateException("Unknown eval result type $this")
+                    }
+                }
+                is ResultWithDiagnostics.Failure -> {
+                    throw ReplCompilerException(resultWithDiagnostics)
+                }
+                else -> throw IllegalStateException("Unknown result")
+            }
+        } finally {
+            isExecuting = false
+        }
+    }
+}
